Added SVGCanvas as alternative to GWTCanvas.
This is mostly useful just for folks trying to paint a large
canvas on Firefox on Windows. There is currently a bug in FF that
prevents large canvasses from being drawn, but you can draw it OK
using SVG.

To use this, generally, an app will have a Firefox-specific override
to substitute a SVGCanvas object for the GWTCanvas object it normally
uses, and everything else works the same.

Also tweaked the gwtcanvas example to build a version using SVGCanvas,
and to test the fillText() method.